python_code/pyproject.toml
[project]
dynamic = ["version"]
name = "python_code"
dependencies = [
"python-pptx",
"Pillow"
]